requirements.txt
tqdm==4.42.1
cycler==0.10.0
foolbox==3.3.1
freeplot==0.0.8
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
torch==1.4.0
torch-tb-profiler==0.1.0
torchvision==0.5.0
numpy==1.18.1
Pillow==8.2.0
scikit_learn==0.24.2
